+++ packages/react-magma-dom/src/components/Combobox/MultiCombobox.test.js
@@ -1,7 +1,9 @@
 import React from 'react';
-import { render, fireEvent } from '@testing-library/react';
+import { render, fireEvent, act } from '@testing-library/react';
 import { Combobox as MultiCombobox } from '.';
 import { magma } from '../../theme/magma';
+import { Modal } from '../Modal';
+import { Button } from '../Button';
 
 describe('MultiCombobox', () => {
   const labelText = 'Label';
@@ -698,6 +700,46 @@ describe('MultiCombobox', () => {
     );
   });
 
+  it('when disabled, selected items should not be removable', () => {
+    const { getByLabelText, getByText } = render(
+      <MultiCombobox
+        isMulti
+        labelText={labelText}
+        items={items}
+        disabled
+        initialSelectedItems={['Red']}
+      />
+    );
+
+    expect(getByLabelText(labelText, { selector: 'input' })).toHaveAttribute(
+      'disabled'
+    );
+    const selectedItem = getByText('Red', { selector: 'button' });
+    expect(selectedItem).toBeInTheDocument();
+    fireEvent.click(selectedItem);
+
+    expect(selectedItem).toBeInTheDocument();
+  });
+
+  it('when disabled, isClearable button should not be clickable', () => {
+    const { getByLabelText, getByText } = render(
+      <MultiCombobox
+        isMulti
+        labelText={labelText}
+        items={items}
+        disabled
+        isClearable
+        initialSelectedItems={['Red']}
+      />
+    );
+
+    expect(getByLabelText(labelText, { selector: 'input' })).toHaveAttribute(
+      'disabled'
+    );
+    fireEvent.click(getByLabelText(/reset selection/i));
+    expect(getByText('Red', { selector: 'button' })).toBeInTheDocument();
+  });
+
   it('should allow a selection to be cleared with isClearable prop', () => {
     const { getByLabelText, getByText, queryByText } = render(
       <MultiCombobox
@@ -1094,6 +1136,91 @@ describe('MultiCombobox', () => {
           inputValue: 'Red',
         })
       );
+    });
+  });
+
+  describe('inside a modal', () => {
+    const items = [
+      { label: 'Red', value: 'red' },
+      { label: 'Blue', value: 'blue' },
+      { label: 'Green', value: 'green' },
+    ];
+
+    beforeEach(() => {
+      jest.useFakeTimers();
+    });
+
+    afterEach(() => {
+      jest.useRealTimers();
+      jest.resetAllMocks();
+    });
+
+    it('should close the menu when escape key is pressed, and retain the active modal', () => {
+      const { getByLabelText, queryByText, getByText } = render(
+        <Modal testId="modal" isOpen>
+          Modal Content
+          <MultiCombobox isMulti labelText={labelText} items={items} />
+        </Modal>
+      );
+
+      const renderedCombobox = getByLabelText(labelText, {
+        selector: 'input',
+      });
+
+      expect(renderedCombobox).toBeInTheDocument();
+
+      fireEvent.click(renderedCombobox);
+
+      expect(getByText(items[0].label)).toBeInTheDocument();
+
+      fireEvent.keyDown(getByLabelText(labelText, { selector: 'input' }), {
+        key: 'Escape',
+        code: 27,
+      });
+
+      expect(
+        queryByText(items[2], { selector: 'Red' })
+      ).not.toBeInTheDocument();
+
+      expect(getByText('Modal Content')).toBeInTheDocument();
+    });
+
+    it('should close the modal on escape after menu is closed on escape', async () => {
+      const onEscKeyMock = jest.fn();
+      const { getByLabelText, queryByText, getByText, debug } = render(
+        <Modal testId="modal" isOpen onEscKeyDown={onEscKeyMock}>
+          Modal Content
+          <MultiCombobox isMulti labelText={labelText} items={items} />
+        </Modal>
+      );
+      const renderedCombobox = getByLabelText(labelText, {
+        selector: 'input',
+      });
+      fireEvent.click(renderedCombobox);
+      expect(getByText(items[0].label)).toBeInTheDocument();
+
+      fireEvent.keyDown(getByLabelText(labelText, { selector: 'input' }), {
+        key: 'Escape',
+        code: 27,
+      });
+
+      expect(
+        queryByText(items[2], { selector: 'Red' })
+      ).not.toBeInTheDocument();
+
+      expect(getByText('Modal Content')).toBeInTheDocument();
+
+      fireEvent.keyDown(getByText('Modal Content'), {
+        key: 'Escape',
+        code: 27,
+      });
+
+      await act(async () => {
+        jest.runAllTimers();
+      });
+
+      expect(onEscKeyMock).toHaveBeenCalled();
+      expect(queryByText('Modal Content')).not.toBeInTheDocument();
     });
   });
 });