+++ packages/react-magma-dom/src/components/Input/index.tsx
@@ -18,6 +18,7 @@ export const Input = React.forwardRef<HTMLInputElement, InputProps>(
       children,
       containerStyle,
       errorMessage,
+      hasCharacterCounter = true,
       helperMessage,
       iconPosition,
       id: defaultId,
@@ -27,6 +28,7 @@ export const Input = React.forwardRef<HTMLInputElement, InputProps>(
       labelStyle,
       labelText,
       labelWidth,
+      maxCount,
       maxLength,
       messageStyle,
       testId,
@@ -37,7 +39,13 @@ export const Input = React.forwardRef<HTMLInputElement, InputProps>(
     const id = useGenerateId(defaultId);
 
     const descriptionId =
-      errorMessage || helperMessage || maxLength ? `${id}__desc` : null;
+      errorMessage || helperMessage || maxCount || maxLength
+        ? `${id}__desc`
+        : null;
+
+    const maxCharacters = typeof maxCount === 'number' ? maxCount : maxLength;
+
+    const maxLengthNum = !hasCharacterCounter && maxLength ? maxLength : undefined;
 
     const isInverse = useIsInverse(props.isInverse);
 
@@ -66,6 +74,7 @@ export const Input = React.forwardRef<HTMLInputElement, InputProps>(
         containerStyle={containerStyle}
         errorMessage={errorMessage}
         fieldId={id}
+        hasCharacterCounter={hasCharacterCounter}
         helperMessage={helperMessage}
         iconPosition={iconPosition}
         isLabelVisuallyHidden={isLabelVisuallyHidden}
@@ -77,6 +86,7 @@ export const Input = React.forwardRef<HTMLInputElement, InputProps>(
         labelText={labelText}
         labelWidth={labelWidth}
         maxLength={maxLength}
+        maxCount={maxCount}
         messageStyle={messageStyle}
         testId={testId && `${testId}-formFieldContainer`}
       >
@@ -86,12 +96,16 @@ export const Input = React.forwardRef<HTMLInputElement, InputProps>(
             descriptionId ? descriptionId : props['aria-describedby']
           }
           aria-invalid={!!errorMessage}
-          hasError={!!errorMessage || characterLength > maxLength}
+          hasError={
+            !!errorMessage ||
+            (hasCharacterCounter && characterLength > maxCharacters)
+          }
           iconPosition={iconPosition}
           id={id}
           inputSize={inputSize}
           inputLength={characterLength}
           isInverse={isInverse}
+          maxLength={maxLengthNum}
           onChange={handleChange}
           onClear={handleClear}
           ref={ref}