+++ packages/react-magma-dom/src/components/Table/Table.stories.tsx
@@ -327,6 +327,7 @@ export const ControlledPagination = args => {
         page={pageIndex}
         rowsPerPage={rowsPerPage}
         isInverse={args.isInverse}
+        hasSquareCorners={args.hasSquareCorners}
       />
     </Card>
   );
@@ -380,6 +381,63 @@ export const UncontrolledPagination = args => {
       />
     </>
   );
+};
+
+export const PaginationWithSquareCorners = args => {
+  const [pageIndex, setPageIndex] = React.useState<number>(1);
+  const [rowsPerPage, setRowsPerPage] = React.useState<number>(10);
+
+  function handlePageChange(_, page) {
+    setPageIndex(page);
+  }
+
+  function handleRowsPerPageChange(newNumberOfRows) {
+    setRowsPerPage(newNumberOfRows);
+  }
+
+  const rowsToShow = rowsLong.slice(
+    (pageIndex - 1) * rowsPerPage,
+    (pageIndex - 1) * rowsPerPage + rowsPerPage
+  );
+
+  return (
+    <div style={{ background: magma.colors.primary600, padding: '20px' }}>
+      <Table style={{ background: magma.colors.neutral100 }} {...args}>
+        <TableHead>
+          <TableRow>
+            <TableHeaderCell>Column</TableHeaderCell>
+            <TableHeaderCell>Column</TableHeaderCell>
+            <TableHeaderCell>Column</TableHeaderCell>
+            <TableHeaderCell>Column</TableHeaderCell>
+          </TableRow>
+        </TableHead>
+        <TableBody>
+          {rowsToShow.map((row, i) => (
+            <TableRow key={`row${i}`}>
+              {row.map((cell, j) => (
+                <TableCell key={`cell${i}_${j}`}>{cell}</TableCell>
+              ))}
+            </TableRow>
+          ))}
+        </TableBody>
+      </Table>
+      <TablePagination
+        itemCount={rowsLong.length}
+        onRowsPerPageChange={handleRowsPerPageChange}
+        onPageChange={handlePageChange}
+        page={pageIndex}
+        rowsPerPage={rowsPerPage}
+        isInverse={args.isInverse}
+        hasSquareCorners={args.hasSquareCorners}
+      />
+    </div>
+  );
+};
+PaginationWithSquareCorners.args = {
+  hasSquareCorners: true,
+  hasHoverStyles: false,
+  hasVerticalBorders: false,
+  hasZebraStripes: false,
 };
 
 export const PaginationInverse = args => {
@@ -764,4 +822,50 @@ export const AdjustableRowNumber = args => {
 };
 AdjustableRowNumber.args = {
   numberRows: 300,
+};
+
+export const NoRowsPerPageControl = args => {
+  const [pageIndex, setPageIndex] = React.useState<number>(1);
+  const rowsPerPage = 5;
+
+  function handlePageChange(_, page) {
+    setPageIndex(page);
+  }
+
+  const rowsToShow = rowsLong.slice(
+    (pageIndex - 1) * rowsPerPage,
+    (pageIndex - 1) * rowsPerPage + rowsPerPage
+  );
+
+  return (
+    <Card isInverse={args.isInverse}>
+      <Table {...args}>
+        <TableHead>
+          <TableRow>
+            <TableHeaderCell>Column</TableHeaderCell>
+            <TableHeaderCell>Column</TableHeaderCell>
+            <TableHeaderCell>Column</TableHeaderCell>
+            <TableHeaderCell>Column</TableHeaderCell>
+          </TableRow>
+        </TableHead>
+        <TableBody>
+          {rowsToShow.map((row, i) => (
+            <TableRow key={`row${i}`}>
+              {row.map((cell, j) => (
+                <TableCell key={`cell${i}_${j}`}>{cell}</TableCell>
+              ))}
+            </TableRow>
+          ))}
+        </TableBody>
+      </Table>
+      <TablePagination
+        itemCount={rowsLong.length}
+        onPageChange={handlePageChange}
+        page={pageIndex}
+        rowsPerPage={rowsPerPage}
+        isInverse={args.isInverse}
+        hasSquareCorners={args.hasSquareCorners}
+      />
+    </Card>
+  );
 };