+++ packages/react-magma-dom/src/components/Tabs/TabPanel.tsx
@@ -12,7 +12,7 @@ const StyledTabPanel = styled.div<{
   color: ${props =>
     props.isInverse
       ? props.theme.colors.neutral100
-      : props.theme.colors.neutral};
+      : props.theme.colors.neutral700};
   font-family: ${props => props.theme.bodyFont};
   flex: 1;
   height: 100%;