+++ packages/react-magma-dom/src/components/Textarea/index.tsx
@@ -26,7 +26,7 @@ export interface TextareaProps
   /**
    * A number value which gives Character Counter the maximum length of allowable characters in an Textarea.
    */
-  maxLength?: number;
+  maxCount?: number;
   /**
    * @internal
    */
@@ -55,6 +55,7 @@ export const Textarea = React.forwardRef<HTMLTextAreaElement, TextareaProps>(
     const {
       containerStyle,
       errorMessage,
+      hasCharacterCounter = true,
       helperMessage,
       id: defaultId,
       isLabelVisuallyHidden,
@@ -62,6 +63,7 @@ export const Textarea = React.forwardRef<HTMLTextAreaElement, TextareaProps>(
       labelStyle,
       labelText,
       labelWidth,
+      maxCount,
       maxLength,
       messageStyle,
       testId,
@@ -73,6 +75,9 @@ export const Textarea = React.forwardRef<HTMLTextAreaElement, TextareaProps>(
 
     const id = useGenerateId(defaultId);
     const descriptionId = errorMessage || helperMessage ? `${id}__desc` : null;
+    const maxCharacters = typeof maxCount === 'number' ? maxCount : maxLength;
+
+    const maxLengthNum = !hasCharacterCounter && maxLength ? maxLength : undefined;
 
     const [value, setValue] = React.useState<
       string | ReadonlyArray<string> | number
@@ -103,6 +108,7 @@ export const Textarea = React.forwardRef<HTMLTextAreaElement, TextareaProps>(
         containerStyle={containerStyle}
         errorMessage={errorMessage}
         fieldId={id}
+        hasCharacterCounter={hasCharacterCounter}
         helperMessage={helperMessage}
         isLabelVisuallyHidden={isLabelVisuallyHidden}
         isInverse={isInverse}
@@ -111,6 +117,7 @@ export const Textarea = React.forwardRef<HTMLTextAreaElement, TextareaProps>(
         labelText={labelText}
         labelPosition={labelPosition}
         labelWidth={labelWidth}
+        maxCount={maxCount}
         maxLength={maxLength}
       >
         <StyledTextArea
@@ -120,8 +127,12 @@ export const Textarea = React.forwardRef<HTMLTextAreaElement, TextareaProps>(
           }
           aria-invalid={!!errorMessage}
           data-testid={testId}
-          hasError={!!errorMessage || characterLength > maxLength}
+          hasError={
+            !!errorMessage ||
+            (hasCharacterCounter && characterLength > maxCharacters)
+          }
           id={id}
+          maxLength={maxLengthNum}
           isInverse={isInverse}
           onChange={handleChange}
           ref={ref}