+++ website/react-magma-docs/src/components/SubPageTabs/index.js
@@ -70,18 +70,7 @@ export const SubPageTabs = ({ pageData, hasHorizontalNav }) => {
 
     window.scrollBy({ top: originalTop, left: 0, behavior: 'smooth' });
 
-    const checkIfDone = setInterval(function () {
-      const atBottom =
-        window.innerHeight + window.pageYOffset >=
-        document.body.offsetHeight - 2;
-      if (distanceToTop(targetAnchor) === 0 || atBottom) {
-        targetAnchor.tabIndex = '-1';
-        targetAnchor.focus();
-        window.history.pushState('', '', '#' + targetID);
-        clearInterval(checkIfDone);
-      }
-    }, 100);
-
+    window.history.pushState('', '', '#' + targetID);
     setActiveTab(index);
   };