+++ patterns/header/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Change Log
 
+## 7.0.0-next.0
+
+### Patch Changes
+
+- Updated dependencies [9a09b8b55]
+  - react-magma-dom@4.2.0-next.0
+
+## 6.0.1
+
+## 6.0.1-next.0
+
+### Patch Changes
+
+- Updated dependencies [871820e8d]
+  - react-magma-dom@4.1.1-next.0
+
+## 6.0.0
+
 ## 6.0.0-next.1
 
 ### Patch Changes