+++ website/react-magma-landing/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## 1.1.2
+
+### Patch Changes
+
+- 1238e8d: - Display latest version based on React compatibility
+  - Remove old versions from list that don't have docs sites
+
+## 1.1.2-next.0
+
+### Patch Changes
+
+- a81b0728b: - Display latest version based on React compatibility
+  - Remove old versions from list that don't have docs sites
+
 ## 1.1.1
 
 ### Patch Changes